change userid column type to match users table

+++ backend/sql/tables/finance_data.sql
@@ -7,7 +7,7 @@ CREATE TABLE FINANCE_DATA(
     tuition INT NOT NULL,
     roomboard INT NOT NULL,
     payperiod VARCHAR(20), -- i.e. fall2025
-    userid SMALLINT NOT NULL,
+    userid INT NOT NULL,
     FOREIGN KEY (userid) REFERENCES USERS(userid)
 );
